chore: update default mariadb params
Change default MariaDB volume name to onyxdb-files

Update dockerfile to set a 24 hours wait_timeout configuration for MariaDB. This fixes the problem where server idles overnight and by the morning the database has dropped the idle connection. All the connecting players then observe their data missing.

Added a comment about importance of the proper password (although check all things security, don't rely on just password) for production servers. In just 2 days the server and database were up for a test launch the database got brute-forced, all the tables dumped and dropped, and a note left in a new table demanding $800 in bitcoin for recovery of the data (a single row with permissions for my testing, lol).