style(tmpl): unify indentation for tmpl files
Change indentation of all the .tmpl files to tabs.
Unification is long due because some files used mixed indentation, some
used 2 spaces, some used 4 spaces and tabs. It was a mess.
Tabs were chosen based on the most common indentation style throughout
the existing files, as well as styles of the other mainstream repos.

Change default file association for .tmpl files to plaintext as pattern
markers make invalid html, which causes errors on automatic validation or
formatting.

This change will be reflected in the .editorconfig in the following
commit.