fix: add debug define
DEBUG define in the DME file allows contributors to use breakpoints with
the langserver integration and run the profiler. Cannot confirm if this
affects the production performance in any way, but will keep the define
set by default as new contributors will get confused about breakpoints
not working in VSCode.