fix: database not reconnecting on timeout
Fix sql_queries crashing on a disconnected database. Adjusting
`wait_timeout` is no longer needed and can be left at the default 8
hours as any failing query will automatically try to re-establish the
connection.

It's important to note that the database code is very outdated, relying
on a Dantom.DB library from many years ago. Being a very outdated code,
`_dm_db_is_connected()` does not actually check for the timed out
database connection and still returns "true" as if the database is still
there.

Refactoring and updating the database core is out of the scope of this
commit.