refactor: mining vendor
Move list of vending equipment into the object itself instead of global.
Historically it was global because there was only one equipment vendor
machine and some items were quite unique and deliverable through a
shuttle only. Right now this behavior is redundant. Althought this
introduces an ability to just reassembe the vending machine to reset
amounts, it's not critical and a todo is added to address this later.
Circuit board is moved to the respective folder.
Some other generic TODO's are added.