feat(mining): pickaxe ore crushing
Add a random effect on attempting to click a piece of ore with pickaxe.
- Ore jostles and backfires into the miner
- Ore gets crushed and disappears
- The valuable part of the ore is destroyed and the ore turns into a
  piece of slag
- The miner strikes the ore just right to get it processed into a random
  product e.g. smelting or compressing
- The miner "strikes it rich" and receives 2-5 pieces of processed
  material
Intended to make pickaxe mining even more annoying, but with a quirk.
It's year 2500+, use drills or heavy duty equipment. Or block incoming
pieces of ore with something. Or just mine carefully.
The feature is a bit experimental in its nature and should be adjusted
as needed after live usage.