feat: disable problematic events
Disable the following events by unticking them from the main .dme file:
- APC damage
- Biohazard outbreak (blob)
- Communications blackout (broken telecomms)
- Electrical storm (exploding lamps and disabled APCs)
- Ion storm (random laws for synthetics)
- Meteor wave
- Mimics invasion
- Space dust (micrometeors at random locations)
- Spontaneous appendicitis
- Xenomorph infestation (specifically larva spawning, facehuggers are
  still enabled)

Most random events received a TODO or multiple TODOs detailing what
needs to be changed as well as why the change is needed.

The reasons provided are my personal observations and may be adjusted
based on feedback from anyone willing to review the commit and research
the system/event in question.

Some events received an mtth (median time to happen, basically a chance
config) adjustment in order to better match current game balance, with
reasons provided in the TODO of the file most of the time.

Replace name-based landmark checking with type-based.

Remove redundant boilerplate condition descriptions stating that event
can only run if it's not already running (duh).