feat: increase max message length
Change default message length limit from 1024 to 2048 characters. As
BYOND and the build itself are much more robust than a decade ago,
there's no reason to be too restrictive given that there are players who
do manage to hit 1024 character limit.
While this is rare, there is no major downsides to this change.