feat: return text to user on failed sanitize
Print user input back to their chat in case sanitization failed.
Solves an annoying issue with users losing their long input text if it
hits the message length limit, as there was no way to know about the
limit ahead of time and no way to get the text back.