feat: make extended require at least one player
Add a threshold of at least one player to the extended gamemode. This in
needed because we have random events that overflow the station if it's
on standby for a long time with no players online, e.g. overnight. The
only solution for players joining after such a delay then is to just ask
for a round restart. This limit ensures that rounds don't start if no
one is ready to play.