feat: change roundend map vote time
Change roundend map vote from arriving to centcomm to 1 minute before
shuttle arrives on station. This is usually the time when most of the
crew is just idling and waiting for departure, but current default of 2
minute vote allows to vote for the map even after boarding the shuttle.

The previous vote was set to trigger on centcomm arrival which
overlapped with the end of round griefing which caused most people to
miss the vote.