fix: load map mapent not accepting files
Fix "load map" map entinty only loading maps if `ev_map_path` is a text
string. In the Example map on a showcase floor map loading didn't work
because map was passed with single quotation marks making it a file.
With this change the map is loaded regardless of the quotation mark
styles.