refactor(code): move separate descriptions to respective object defines
Move all separated descriptions and examine logic from the
code/modules/examine/descriptions/ folder back to the defines of the
respective objects.

This removes duplicate object definitions that are valid BYOND code, but
grab attention of the DM language client and force "Jump to source"
feature to navigate to these descriptions, rather than showing the
actual definition of the object.