feat(code): add support for non-list firemodes
Currently firemodes are defined on guns as lists of variables that will
be changed on mode switch. However, there's a /datum/firemode type that
these lists get implicitly converted to by the base /obj/item/gun
initializer.

This commit adds a check for the list type to the gun initializer.
Converting the entry from list of settings into a firemode happens only
if it is necessary.
This allows to pass and reuse firemodes directly instead of using lists
of settings and relying on implicit conversion.