fix(code): duplicate sanitize call on base living mob type
Fix sanitize() being called for the second time in /mob/living/say,
which most subtypes call back to in their overrides while sanitizing on
their own.

This fixes escapable characters appearing in their html_encoded form,
e.g. double quote `"` appearing as `&#34;` or single quote `'` appearing
as `&#39;`.

The duplicate was introduced in the following commit as part of the say
proc refactoring, but is redundant as all the subtypes already sanitize
their input.
https://github.com/ChaoticOnyx/OnyxBay/commit/2a91be07c2a45a40b86175e82f09d5ba92349f43#diff-b11078c1ae5a0caf0e47a062537a4b7672f7090049dc8452198ae9140d7a8541R139

The only non-sanitized say() was /mob/living/deity, however it is
currently not used in any live gamemodes and hasn't been maintained for
a few years. For the purposes of code maintenance a sanitize call was
added to it.