style(code): standardize multiline spans into single line style
Multiline or string composition spans brought to a common span-per-line
style, with some places receiving temporary variables to store text that
is then wrapped with a span.

Note that telecomms broadcaster is an explicit exception to this styling
as radio messages are dynamically built and there's no clear way to
apply spans to them.