style(code): standardize string-embedded spans into SPAN macros
Span is considered string-embedded if it is part of a string that has
some additional content besides the span element.

That means that if there's at least one other character between the
<span...></span> and surrounding it quotation marks, the span is
string-embedded.
Example:
  msg += "That's a <span class='danger>DONUT</span>"

All string-embedded spans were converted to SPAN() macros according to
the codebase standard or turned into non-string-embedded spans if extra
contents of the string was just whitespace.