style(code): standardize all string spans to SPAN() macro style
Following the established codebase standard, all text style spans are
replaced with SPAN("class", "text") macro.

This includes fixing broken or unmatched tags and non-standard spans,
e.g. ones that don't have a class defined, but rather change the style.

Spans that are part of paper tag processing, regexes, satellite
broadcaster code, and spans that define an id are untouched as an
exception to the codestyle standard, as macro is not applicable to them.