docs: add "unchecked information" notice to the contributing guide
Add a notice that contributing guide was inherited from OnyxBay, and as
such might be outdated or not relevant to ZeroOnyx's standards.

Also, turns out there was a stealthy "penis" in the guide to
contributing. Not the place for it. Contribution guide should be neutral
and welcoming to newcomers, not immature.

And donuts make a better example anyways.