refactor(code): remove redundant return values from mouse drop code
MouseDrop code is called by BYOND when something is drag-n-dropped.
BYOND does not react in any way to the return value, so it's redundant.

MouseDrop_T proc is spawned when base atom's implementation of MouseDrop
is called, but also doesn't process the return value in any way.

All return values of MouseDrop and MouseDrop_T procs were removed to
improve readability and reduce potential confusion for future
contributors.