refactor(code): move cooked_types var back to the food item definition
When variables of an object type are split across multiple files,
additional definitions override the type tree path and "Jump to Source"
is only able to jump to the latest file it parsed. This makes code
navigation harder.

On top of that, defining variables of a type in multiple files may lead
to confusion as contributors might not expect multiple definition
sources.
Unifying definitions allows a single source of truth define which is
much more clean and maintainable.