fix(code): surgery interactions not resolving the click
Successful surgery interactions should return TRUE in the `attackby()` code, thus resolving the "attack"/click and telling click handling that `afterattack()` should not be called.
Return value on the surgery condition was incorrectly set to FALSE.

This caused double interactions for items that both act as a surgery
tool and have an `afterattack` action, e.g. syringes.