fix(code): duplicate item visibility when removing embedded objects
When items are embedded in a mob, they are added to the mob's `contents`
list. This automatically adds these items to the `client.screen` var
which is intended to make items in the mob's inventory visible on the
screen HUD.

However, while inventory items are remobed from the mob using the
`drop()` proc which handles removal from the list of visible objects
among other cleanup, items that are embedded into bodyparts (e.g. after
being thrown) were forcibly removed with `forceMove()` call as they
didn't naturally fit the `drop()` logic designed for inventory
interactions. This simply updated the item's location without much
consideration for anything else.

The behavior caused "ghost" items to show up in HUD of the user if their
mob had an item embedded in them and someone else "yanked out" the
embedded item into their hands. The only way to get rid of the item
being drawn over the user's inventory was for the user to pick it up and
drop or give it to someone else so the HUD visibility could be properly
reset.

Forcible location change of the embedded item was replaced with a
`drop()` call with `force` parameter applied which allows item to be
properly cleaned up from the screens and reset into it's normal state.
The `force` parameter allows to bypass normal inventory checks which
would fail otherwise as embedded item isn't actually in any of the mob's
inventory slots.