feat(code): remove the delay required to put on the RIG control module
RIGs already have a hefty delay on seal adjustments and
activation/deactivation.
Adding additional delay to putting on the control module which is hardly
more than a backpack is redundant.

Additionally, dropping or `forceMove()`ing RIG module on unsuccessful
equip attempt caused visual glitches where HUD plane and layer would get
applied even after the item is dropped.
The reason behind dropping the module was to deal with the fact that it
technically was on the back of the user even before the delay, just not
visually, and so "failed" equip attempt couldn't put it back in hands
because of the complicated call chain.