fix(code): gun warcrime handler not typechecking the grab
When using a gun on someone else while targeting their mouth, a so
called "warcrime" interaction happens which allows to execute people.
The `handle_war_crime()` proc checks if user is grabbing the target with
their inactive hand, but didn't check if the result was actually a grab
which resulted in a runtime.

This was fixed with an explicit typecheck, and redundant checked access
operators (e.g. `G?.affecting`) were removed with normal access.