fix(code): storage subtypes not returning the removal result
The return value of the `remove_from_storage` proc is used by the item
interactions code to determine if item was successfully removed or
stayed inside the storage.

Subtypes of storage had inconsistent overrides of the
`remove_from_storage` proc which led to interactions failing midway and
dropping items on the floor instead of letting the user take them out of
storage, e.g. when cigars were taken from the cigar case.