feat(code): make escape pods eject even if main shuttle was forced
It makes more sense for the escape pods to eject together with the
shuttle regardless of the departure type, scheduled or forced.

The previous behavior of pods just standing by and not receiving a
launch signal if shuttle was forcibly launched was confusing to the
players and not worth keeping.