fix(code): storage insertion not handling null user
Certain parts of the code attempt to insert items into containers
without direct user input for the action. This results in calls to
storage procs with `user` param set to null.

Intent checking for certain item exceptions was updated to null-checked
access to the `user` variable in order to support userless scenarios.