fix(code): uniform accessories not equipping and disappearing
Due to uniform accessories not having a real slot and not being equipped
anywhere during the character creation on roundstart, the result of
`can_unequip()` proc for them was `false`, which resulted in them not
being droppable, which in turn caused them to get stuck in mob's conents
with no valid inventory reference.

Accessories are now force-dropped on being attached to a uniform, they
don't have a real slot that would make them non-unequippable and should
always be attachable if accessory checks pass.