fix(code): crayon drawings not rendering immediately after spawning
Replace reliance on newly generated icons for crayon drawings with a
combination of regular icon rendering, color matrix transforms and a
shading image overlay.

Old flow relied on generating new icons at runtime that had to be
uploaded to the client caches, which might not happen in time and leave
the drawing blank until client attempts to refetch or redraw the icon
(e.g. with a reconnect).

New flow reuses same base icon files and utilizes client transforms to
modify the base drawing color and shading overlay, basically
guaranteeing that clients will have needed icon in cache when drawing is
created.