chore: update maximum size of client upload to 20MiB
There is no heavy load on the server right now and players who upload
music to the custom tapes in the bar sometimes hit the limit.
While compressing the music is an option, it's annoying extra steps.

For the current context it makes more sense to raise the file size limit
even if it causes some spontaneous one-time stutters for players
receiving the file.